Create README.md
About
This application is for a third party company to be able to add event details to their database. Details can range from customers, events speakers and much more .

Setup.
To get started you will need to clone this repo and the client side located at
shane86920915/EventPlanner (github.com)

Server side setup
Once cloned, ensure Visual Studio is installed and open the cloned solution in Visual Studio. It is recommended to clean and build the solution to fix potential build errors.
Ensure the startup project is selected and run the project. The project will open in a web browser with the URL set to the local host. After registration is completed, customers, speakers, and sponsors can be created.

Features
Features include crud functionality for speakers, customers, and sponsors.
there is a search bar for all entities, and also an api for gps purposes.